+++ services/maps/mgmt/2017-01-01/maps/client.go
@@ -1,6 +1,6 @@
 // Package maps implements the Azure ARM Maps service API version 2017-01-01-preview.
 //
-// Resource Provider
+// Azure Maps
 package maps
 
 // Copyright (c) Microsoft Corporation. All rights reserved.