+++ services/maps/mgmt/2017-01-01/maps/mapsapi/interfaces.go
@@ -8,7 +8,7 @@ package mapsapi
 
 import (
 	"context"
-	"github.com/Azure/azure-sdk-for-go/services/maps/mgmt/2017-01-01-preview/maps"
+	"github.com/Azure/azure-sdk-for-go/services/maps/mgmt/2017-01-01/maps"
 	"github.com/Azure/go-autorest/autorest"
 )