+++ services/preview/maps/mgmt/2020-02-01-preview/maps/maps.go
@@ -14,7 +14,7 @@ import (
 	"net/http"
 )
 
-// Client is the resource Provider
+// Client is the azure Maps
 type Client struct {
 	BaseClient
 }